added ars-usd.csv and HousePrices.ipynb
ars-usd.csv is a date and conversion rate between ARS and USD for the relevant period. HousePrices.ipynb is a notebook with some data cleaning and graphing that requires another csv, ar_properties_slim.csv, to run. Unfortunately that file is too large (163 MB, GitHub limit is 25 MB), so I will store it elsewhere.